+++ components/MapView/index.js
@@ -14,15 +14,10 @@ const Maps = ({ organization }) => {
     let isSubscribed = true;
 
     async function fetchRegion() {
-      await getData('assetMapRegion').then((data) => {
+      await getData('homeMapRegion').then((data) => {
         if (isSubscribed) {
           if (!data) {
-            setRegion({
-              latitude: 18.4861,
-              longitude: -69.9312,
-              latitudeDelta: 0.0922,
-              longitudeDelta: 0.0421,
-            });
+            handleLocation();
           } else {
             setRegion(data);
           }
@@ -54,7 +49,7 @@ const Maps = ({ organization }) => {
         longitudeDelta: 0.0421,
         latitude,
         longitude,
-      }, 'assetMapRegion');
+      }, 'homeMapRegion');
     }).catch((e) => {
       console.log(e) //eslint-disable-line
     });