+++ domains/DataCollection/Assets/NewAssets/AssetSupplementary/index.js
@@ -11,6 +11,7 @@ import PaperInputPicker from '../../../../../components/FormikFields/PaperInputP
 import PopupError from '../../../../../components/PopupError';
 import { getData } from '../../../../../modules/async-storage';
 import { postSupplementaryAssetForm } from '../../../../../modules/cached-resources';
+import { storeAppVersion } from '../../../../../modules/cached-resources/populate-cache';
 import I18n from '../../../../../modules/i18n';
 import { theme } from '../../../../../modules/theme';
 import { isEmpty } from '../../../../../modules/utils';
@@ -45,7 +46,7 @@ const AssetSupplementary = ({
           const user = await getData('currentUser');
 
           const surveyingUserFailSafe = await surveyingUserFailsafe(user, surveyingUser, isEmpty);
-          const appVersion = await getData('appVersion') || '';
+          const appVersion = await storeAppVersion() || '';
 
           let formObjectUpdated = addSelectTextInputs(values, formObject);
           formObjectUpdated = cleanLoopSubmissions(values, formObjectUpdated);
@@ -134,7 +135,7 @@ const AssetSupplementary = ({
                   <PaperButton
                     disabled={!validForm()}
                     style={{ backgroundColor: validForm() ? 'green' : '#f75231' }}
-                    onPressEvent={() => formikProps.handleSubmit()}
+                    onPress={() => formikProps.handleSubmit()}
                     icon={validForm() ? 'plus' : 'alert-octagon'}
                     buttonText={validForm() ? I18n.t('global.submit') : I18n.t('assetForms.attachForm')}
                   />
@@ -142,7 +143,7 @@ const AssetSupplementary = ({
                 <PaperButton
                   mode="text"
                   buttonText={I18n.t('assetCore.tapCreateAsset')}
-                  onPressEvent={() => setPage('assetCore')}
+                  onPress={() => setPage('assetCore')}
                 />
               </View>
               <PopupError