+++ domains/DataCollection/Forms/IdentificationForm/index.js
@@ -15,6 +15,7 @@ import yupValidationPicker from '../../../../components/FormikFields/YupValidati
 import PopupError from '../../../../components/PopupError';
 import { getData } from '../../../../modules/async-storage';
 import { postIdentificationForm } from '../../../../modules/cached-resources';
+import { storeAppVersion } from '../../../../modules/cached-resources/populate-cache';
 import I18n from '../../../../modules/i18n';
 import { layout, theme } from '../../../../modules/theme';
 import { isEmpty } from '../../../../modules/utils';
@@ -52,7 +53,8 @@ const IdentificationForm = ({
 
             formObject.surveyingOrganization = surveyingOrganization || user.organization;
             formObject.surveyingUser = await surveyingUserFailsafe(user, surveyingUser, isEmpty);
-            formObject.appVersion = await getData('appVersion') || '';
+
+            formObject.appVersion = await storeAppVersion() || '';
             formObject.phoneOS = Platform.OS || '';
 
             formObject.latitude = values.location?.latitude || 0;
@@ -128,7 +130,7 @@ const IdentificationForm = ({
                 />
               ) : (
                 <PaperButton
-                  onPressEvent={formikProps.handleSubmit}
+                  onPress={formikProps.handleSubmit}
                   buttonText={_.isEmpty(formikProps.values) ? I18n.t('global.emptyForm') : I18n.t('global.submit')}
                   icon={_.isEmpty(formikProps.values) ? 'alert-octagon' : 'plus'}
                   style={{ backgroundColor: _.isEmpty(formikProps.values) ? 'red' : 'green' }}