+++ modules/cached-resources/Post/post.js
@@ -6,36 +6,34 @@ import {
 import checkOnlineStatus from '../../offline';
 import { generateRandomID } from '../../utils';
 
-function postIdentificationForm(postParams) {
-  return new Promise((resolve, reject) => {
-    checkOnlineStatus().then((connected) => {
-      if (connected) {
-        postObjectsToClass(postParams).then((surveyee) => {
-          const surveyeeSanitized = JSON.parse(JSON.stringify(surveyee));
-          resolve(surveyeeSanitized);
-        }, (error) => {
-          reject(error);
-        });
-      } else {
-        getData('offlineIDForms').then(async (idForms) => {
-          const id = `PatientID-${generateRandomID()}`;
-          const idParams = postParams;
-          idParams.localObject.objectId = id;
-          if (idForms !== null || idForms === []) {
-            const forms = idForms.concat(idParams);
-            await storeData(forms, 'offlineIDForms');
-            resolve(idParams.localObject);
-          } else {
-            const idData = [idParams];
-            // idData[id] = postParams;
-            await storeData(idData, 'offlineIDForms');
-            resolve(idParams.localObject);
-          }
-        });
-      }
-    });
+const postIdentificationForm = async (postParams) => {
+  const isConnected = await checkOnlineStatus();
+  if (isConnected) {
+    return postObjectsToClass(postParams).then((surveyee) => {
+      const surveyeeSanitized = JSON.parse(JSON.stringify(surveyee));
+      return surveyeeSanitized;
+    }).catch((error) => error);
+  }
+
+  return getData('offlineIDForms').then(async (offlineIDForms) => {
+    const offlineResidentIdForms = offlineIDForms;
+
+    const idParams = postParams;
+    const { localObject } = idParams;
+
+    localObject.objectId = `PatientID-${generateRandomID()}`;
+
+    if (offlineResidentIdForms) {
+      const forms = offlineResidentIdForms.concat(idParams);
+      await storeData(forms, 'offlineIDForms');
+      return localObject;
+    }
+
+    const idData = [idParams];
+    await storeData(idData, 'offlineIDForms');
+    return localObject;
   });
-}
+};
 
 /** ***********************************************
  * Function to post asset id form offline offline
@@ -47,61 +45,46 @@ function postIdentificationForm(postParams) {
  *
  *********************************************** */
 
-function postAssetForm(postParams) {
-  return new Promise((resolve, reject) => {
-    checkOnlineStatus().then((connected) => {
-      if (connected) {
-        postObjectsToClass(postParams).then((asset) => {
-          const assetSanitized = JSON.parse(JSON.stringify(asset));
-          resolve(assetSanitized);
-        }, (error) => {
-          reject(error);
-        });
-      } else {
-        getData('offlineAssetIDForms').then(async (assetIdForms) => {
-          const id = `AssetID-${generateRandomID()}`;
-          const assetIdParams = postParams;
-          assetIdParams.localObject.objectId = id;
-          if (assetIdForms !== null || assetIdForms === []) {
-            const forms = assetIdForms.concat(assetIdParams);
-            await storeData(forms, 'offlineAssetIDForms');
-            resolve(assetIdParams.localObject);
-          } else {
-            const idData = [assetIdParams];
-            await storeData(idData, 'offlineAssetIDForms');
-            resolve(assetIdParams.localObject);
-          }
-        });
-      }
-    });
-  });
-}
+const postAssetForm = async (postParams) => {
+  const isConnected = await checkOnlineStatus();
+  if (isConnected) {
+    return postObjectsToClass(postParams).then((asset) => {
+      const assetSanitized = JSON.parse(JSON.stringify(asset));
+      return assetSanitized;
+    }).catch((error) => error);
+  }
+  return getData('offlineAssetIDForms').then(async (offlineData) => {
+    const id = `AssetID-${generateRandomID()}`;
+    const assetIdParams = postParams;
+    const offlineAssetForms = offlineData;
+    assetIdParams.localObject.objectId = id;
 
-function postSupplementaryForm(postParams) {
-  return new Promise((resolve, reject) => {
-    checkOnlineStatus().then((connected) => {
-      if (connected && !postParams.parseParentClassID.includes('PatientID-')) {
-        postObjectsToClassWithRelation(postParams).then(() => {
-          resolve('success');
-        }, (error) => {
-          reject(error);
-        });
-      } else {
-        getData('offlineSupForms').then(async (supForms) => {
-          if (supForms !== null) {
-            const forms = supForms.concat(postParams);
-            await storeData(forms, 'offlineSupForms');
-            resolve('success');
-          } else {
-            const supData = [postParams];
-            await storeData(supData, 'offlineSupForms');
-            resolve('success');
-          }
-        });
-      }
-    });
+    if (offlineAssetForms) {
+      const forms = offlineAssetForms.concat(assetIdParams);
+      return storeData(forms, 'offlineAssetIDForms');
+    }
+
+    const idData = [assetIdParams];
+    const storedData = await storeData(idData, 'offlineAssetIDForms');
+    return storedData;
   });
-}
+};
+
+const postSupplementaryForm = async (postParams) => {
+  const isConnected = await checkOnlineStatus();
+  if (isConnected && !postParams?.parseParentClassID?.includes('PatientID-')) {
+    return postObjectsToClassWithRelation(postParams);
+  }
+
+  return getData('offlineSupForms').then(async (supForms) => {
+    if (supForms) {
+      const forms = supForms.concat(postParams);
+      return storeData(forms, 'offlineSupForms');
+    }
+    const supData = [postParams];
+    return storeData(supData, 'offlineSupForms');
+  });
+};
 
 /** ***********************************************
  * Function to post asset supplementary form offline
@@ -112,96 +95,55 @@ function postSupplementaryForm(postParams) {
  * @param {Object} postParams Object normally configured for for Parse-Server Cloud Code
  *
  *********************************************** */
-function postSupplementaryAssetForm(postParams) {
-  return new Promise((resolve, reject) => {
-    checkOnlineStatus().then((connected) => {
-      if (connected && !postParams.parseParentClassID.includes('AssetID-')) {
-        postObjectsToClassWithRelation(postParams).then(() => {
-          resolve('success');
-        }, (error) => {
-          reject(error);
-        });
-      } else {
-        getData('offlineAssetSupForms').then(async (supForms) => {
-          if (supForms !== null) {
-            const forms = supForms.concat(postParams);
-            await storeData(forms, 'offlineAssetSupForms');
-            resolve('success');
-          } else {
-            const supData = [postParams];
-            await storeData(supData, 'offlineAssetSupForms');
-            resolve('success');
-          }
-        });
-      }
-    });
-  });
-}
+const postSupplementaryAssetForm = async (postParams) => {
+  const isConnected = await checkOnlineStatus();
 
-function postHousehold(postParams) {
-  return new Promise((resolve, reject) => {
-    checkOnlineStatus().then((connected) => {
-      if (connected) {
-        postObjectsToClass(postParams).then((result) => {
-          resolve(result.id);
-        }, (error) => {
-          reject(error);
-        });
-      } else {
-        getData('offlineHouseholds').then(async (households) => {
-          const id = `Household-${generateRandomID()}`;
-          const householdParams = postParams;
-          householdParams.localObject.objectId = id;
-          if (households !== null || households === []) {
-            const forms = households.concat(householdParams);
-            await storeData(forms, 'offlineHouseholds');
-            resolve(id);
-          } else {
-            const householdData = [householdParams];
-            // idData[id] = postParams;
-            await storeData(householdData, 'offlineHouseholds');
-            resolve(id);
-          }
-        });
-      }
-    });
+  if (isConnected && !postParams.parseParentClassID.includes('AssetID-')) {
+    return postObjectsToClassWithRelation(postParams);
+  }
+  return getData('offlineAssetSupForms').then(async (supForms) => {
+    if (supForms) {
+      const forms = supForms.concat(postParams);
+      return storeData(forms, 'offlineAssetSupForms');
+    }
+    const supData = [postParams];
+    return storeData(supData, 'offlineAssetSupForms');
   });
-}
+};
 
-function postHouseholdWithRelation(postParams) {
-  return new Promise((resolve, reject) => {
-    checkOnlineStatus().then((connected) => {
-      if (connected) {
-        postObjectsToClassWithRelation(postParams).then((result) => {
-          resolve(result.id);
-        }, (error) => {
-          reject(error);
-        });
-      } else {
-        getData('offlineHouseholdsRelation').then(async (householdsRelation) => {
-          const id = `Household-${generateRandomID()}`;
-          const householdParams = postParams;
-          householdParams.localObject.objectId = id;
-          if (householdsRelation !== null || householdsRelation === []) {
-            const forms = householdsRelation.concat(householdParams);
-            await storeData(forms, 'offlineHouseholdsRelation');
-            resolve(id);
-          } else {
-            const householdData = [householdParams];
-            // idData[id] = postParams;
-            await storeData(householdData, 'offlineHouseholdsRelation');
-            resolve(id);
-          }
-        });
-      }
-    });
+/**
+ * Function to post household form. Used for creating a new household
+ * @param {*} postParams
+ * @returns Househould object
+ */
+const postHousehold = async (postParams) => {
+  const isConnected = await checkOnlineStatus();
+
+  if (isConnected) {
+    return postObjectsToClass(postParams).then((result) => result.id).catch((error) => error);
+  }
+
+  return getData('offlineHouseholds').then(async (offlineHouseholds) => {
+    const households = offlineHouseholds;
+    const householdParams = postParams;
+
+    const { localObject } = householdParams;
+    localObject.objectId = `Household-${generateRandomID()}`;
+
+    if (households) {
+      const forms = households.concat(householdParams);
+      await storeData(forms, 'offlineHouseholds');
+      return forms;
+    }
+    const householdData = [householdParams];
+    await storeData(householdData, 'offlineHouseholds');
+    return householdData;
   });
-}
+};
 
 export {
   postAssetForm,
   postHousehold,
-  postHouseholdWithRelation,
   postIdentificationForm,
   postSupplementaryAssetForm,
   postSupplementaryForm